% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ves_utilities.R
\name{apparent_resistivities}
\alias{apparent_resistivities}
\title{apparent_resistivities}
\usage{
apparent_resistivities(rho, thick, filt, spacing)
}
\arguments{
\item{rho}{A numeric vector with the values of the real resistivities of an Earth model}

\item{thick}{A numeric vector with the values of the thicknesses of the Earth model}

\item{filt}{A numeric vector with the values of the coefficients of the linear filter for an Schlumberger array}

\item{spacing}{A numeric vector with the values of the AB/2 spacing.}
}
\value{
A list with the following elements:This section includes all the functions
\itemize{
\item appres: A numeric vector with the values of the apparent resistivities
\item ab2: A numeric vector with the values of the electrode spacing
}
}
\description{
Function to calculate the apparent resistivities using a given Earth model.
}
\seealso{
Other utilities functions: \code{\link{apparent_resistivities_simple}},
  \code{\link{convolution}}, \code{\link{krtrans}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
