% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibrate}
\alias{calibrate}
\title{calibrate}
\usage{
calibrate(ves, opt.method = c("L-BFGS-B", "SA", "GA", "PSO", "DE"),
  obj.fn = c("rss", "mnad", "mxad", "log_rss", "log_mnad", "log_mxad"),
  par0 = par0, lower = lower, upper = upper, control.par = NULL)
}
\arguments{
\item{ves}{A VES object}

\item{opt.method}{A character string specifying the optimization method used to estimate the real resisitivity.
Currently the following methods are supported:
\itemize{
\item L-BFGS-B: Limited-memory modification of the BFGS quasi-Newton method (optim package)
\item SA: Simulated Annealing (GenSA package)
\item GA: Genetic Algorithms (GA package)
\item PSO: Particle Swarm Optimization (pso package)
\item DE: Differential Evoluation (DEoptim package)
}}

\item{obj.fn}{Objective function used in the parameter estimation}

\item{par0}{A numeric vector with the Initial solution}

\item{lower}{A numeric vector of length equal to the number of layers with the min values of the parameter space}

\item{upper}{A numeric vector of length equal to the number of layers with the max values of the parameter space}

\item{control.par}{A list with the parameters of the optimization algorithm}
}
\value{
This function returns a list with the following entries:
\itemize{
\item par: vector with all parameters
\item value: value of the objective function
\item rho: Numeric vector with the real resistivities
\item thickness: Numeric vector with the layer thicknesses
\item rel.error: The value of the relative error
}
}
\description{
Function to estimate the resistivities and thicnesses of the layers.
}
\seealso{
Other calibration functions: \code{\link{calibrate_ilsqp}},
  \code{\link{calibrate_joint_nls}},
  \code{\link{calibrate_nls}},
  \code{\link{calibrate_seq_joint_nls}},
  \code{\link{calibrate_step_nls}},
  \code{\link{calibrate_step}},
  \code{\link{calibrate_svd}},
  \code{\link{log_mnad_resistivity}},
  \code{\link{log_mxad_resistivity}},
  \code{\link{log_rss_resistivity}},
  \code{\link{mnad_resistivity}},
  \code{\link{mxad_resistivity}},
  \code{\link{relative_error_resitivity}},
  \code{\link{rss_resistivity}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
