% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibrate_joint_nls}
\alias{calibrate_joint_nls}
\title{calibrate_joint_nls}
\usage{
calibrate_joint_nls(ves.list, par0, iterations = 100, ireport = 10)
}
\arguments{
\item{ves.list}{A list of VES objects}

\item{par0}{A numeric vector with the initial model}

\item{iterations}{An integer specifying the maximum number of iterations}

\item{ireport}{An integer specifying the report interval}
}
\value{
This function returns a list with the following entries:
\itemize{
\item par: A numeric vector with the values of the layer resistivities and thicknesses
\item value: The value or the RSS (Residual sum of squares)
\item rel.error: The value of the relative error (in percentage)
\item cal.error: A matrix with the RSS and relative error at each iteration
}
}
\description{
Parameter estimation of more than one VES in a single step.
}
\seealso{
Other calibration functions: \code{\link{calibrate_ilsqp}},
  \code{\link{calibrate_nls}},
  \code{\link{calibrate_seq_nls}},
  \code{\link{calibrate_seq}}, \code{\link{calibrate_svd}},
  \code{\link{calibrate}},
  \code{\link{log_mnad_resistivity}},
  \code{\link{log_mxad_resistivity}},
  \code{\link{log_rss_resistivity}},
  \code{\link{mnad_resistivity}},
  \code{\link{mxad_resistivity}},
  \code{\link{relative_error_resitivity}},
  \code{\link{rss_resistivity}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
