% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibrate_nls}
\alias{calibrate_nls}
\title{calibrate_nls}
\usage{
calibrate_nls(ves, par0, iterations = 100, ireport = 10)
}
\arguments{
\item{ves}{A ves object}

\item{par0}{A numeric vector with the values of the layer resistivities and thicknesses}

\item{iterations}{Number of iterations}

\item{ireport}{Number of iterations to report results on the console}
}
\value{
A list with the following entries:
\itemize{
\item par: A numeric vector with the values of the layer resistivities and thicknesses
\item value: The value or the RSS (Residual sum of squares)
\item rel.error: The value of the relative error (in percentage)
\item cal.error: A matrix with the RSS and relative error at each iteration
\item hessian: First order approximation of the Hessian matrix. This is calculate using the Jacobian matrix.
}
}
\description{
Function to estimate the real resistivities and thicknesses using the nonlinear least-squares approach
proposed by Roy(1999).
}
\references{
Roy, I. An efficient non-linear least-squares 1D inversion scheme for resistivity
and IP sounding data. 1999. Geophysical Prospecting, 47, 4, 527-550.
}
\seealso{
Other calibration functions: \code{\link{calibrate_ilsqp}},
  \code{\link{calibrate_joint_nls}},
  \code{\link{calibrate_seq_joint_nls}},
  \code{\link{calibrate_seq_nls}},
  \code{\link{calibrate_seq}}, \code{\link{calibrate_svd}},
  \code{\link{calibrate}},
  \code{\link{log_mnad_resistivity}},
  \code{\link{log_mxad_resistivity}},
  \code{\link{log_rss_resistivity}},
  \code{\link{mnad_resistivity}},
  \code{\link{mxad_resistivity}},
  \code{\link{relative_error_resitivity}},
  \code{\link{rss_resistivity}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
