% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibrate_seq}
\alias{calibrate_seq}
\title{calibrate_seq}
\usage{
calibrate_seq(ves, opt.method, max.layers = 10, lower = 1, upper = 500)
}
\arguments{
\item{ves}{A VES object}

\item{opt.method}{A character string specifying the optimization method to be used}

\item{max.layers}{An integer with the maximum number of layers to be testsed}

\item{lower}{A numeric value}

\item{upper}{A numeric value}
}
\value{
This function returns a list with the following entries:
\itemize{
\item par: A numeric vector with the values of the layer resistivities and thicknesses
\item value: The value or the RSS (Residual sum of squares)
\item rel.error: The value of the relative error (in percentage)
\item cal.error: A matrix with the RSS and relative error at each iteration
}
}
\description{
Sequential VES inversion using different optimization algorithms
}
\seealso{
Other calibration functions: \code{\link{calibrate_ilsqp}},
  \code{\link{calibrate_joint_nls}},
  \code{\link{calibrate_nls}},
  \code{\link{calibrate_seq_nls}},
  \code{\link{calibrate_svd}}, \code{\link{calibrate}},
  \code{\link{log_mnad_resistivity}},
  \code{\link{log_mxad_resistivity}},
  \code{\link{log_rss_resistivity}},
  \code{\link{mnad_resistivity}},
  \code{\link{mxad_resistivity}},
  \code{\link{relative_error_resitivity}},
  \code{\link{rss_resistivity}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
