% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibrate_step_nls}
\alias{calibrate_step_nls}
\title{calibrate_step_nls}
\usage{
calibrate_step_nls(ves, iterations = 30, ireport = 10, max.layers = 10,
  select.measure = "rss")
}
\arguments{
\item{ves}{A VES object}

\item{iterations}{An integer specifying the maximum number of iterations}

\item{ireport}{An integer specifying the report interval}

\item{max.layers}{An integer specifying the maximum number of layers to include in the
sequential estimation.}

\item{select.measure}{A character string specifying the criteria used to select the best
model from the set of models tried during the sequential estimation. Available options
are:
\itemize{
\item rss: (default) sum of residual squares
\item aic: Akaike Information Criteria
\item bic; Bayesian Information Criteria
}}
}
\value{
This function returns a list with the following entries:
\itemize{
\item par: A numeric vector with the values of the layer resistivities and thicknesses
\item value: The value or the RSS (ResiduSpatial Data Analysis in Ecology and Agricultureal sum of squares)
\item rel.error: The value of the relative error (in percentage)
\item cal.error: A matrix with the RSS and relative error at each iteration
}
}
\description{
Stepwise estimation of true resistivities and thicknesses of a VES.
}
\seealso{
Other calibration functions: \code{\link{calibrate_ilsqp}},
  \code{\link{calibrate_joint_nls}},
  \code{\link{calibrate_nls}},
  \code{\link{calibrate_seq_joint_nls}},
  \code{\link{calibrate_step}},
  \code{\link{calibrate_svd}}, \code{\link{calibrate}},
  \code{\link{log_mnad_resistivity}},
  \code{\link{log_mxad_resistivity}},
  \code{\link{log_rss_resistivity}},
  \code{\link{mnad_resistivity}},
  \code{\link{mxad_resistivity}},
  \code{\link{relative_error_resitivity}},
  \code{\link{rss_resistivity}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
