% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibrate_svd}
\alias{calibrate_svd}
\title{calibrate_svd}
\usage{
calibrate_svd(ves, par0, iterations = 100, ireport = 10)
}
\arguments{
\item{ves}{A vertical electrical sounding object}

\item{par0}{A numeric vector with the initial values of resistivities and thicknesses}

\item{iterations}{An integer specifying the maximum number of iterations}

\item{ireport}{An integer specifying the report interval}
}
\value{
This function returns a list with the following entries:
\itemize{
\item par: A numeric vector with the values of the layer resistivities and thicknesses
\item value: The value or the RSS (Residual sum of squares)
\item rel.error: The value of the relative error (in percentage)
\item cal.error: A matrix with the RSS and relative error at each iteration
}
}
\description{
Function to estimate the layer parameters using the approach proposed by Meju (1992)
}
\references{
Meju, M. An effective ridge regression procedure for resistivity data inversion.
Computers and Geosciences, 18(2-3), 99-119, 1992.
}
\seealso{
Other calibration functions: \code{\link{calibrate_ilsqp}},
  \code{\link{calibrate_nls}},
  \code{\link{calibrate_seq_nls}}, \code{\link{calibrate}},
  \code{\link{log_mnad_resistivity}},
  \code{\link{log_mxad_resistivity}},
  \code{\link{log_rss_resistivity}},
  \code{\link{mnad_resistivity}},
  \code{\link{mxad_resistivity}},
  \code{\link{relative_error_resitivity}},
  \code{\link{rss_resistivity}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
