% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibration_ilsqp}
\alias{calibration_ilsqp}
\title{calibration_ilsqp}
\usage{
calibration_ilsqp(ves, iterations = 100, ireport = 10)
}
\arguments{
\item{ves}{A VES object}

\item{iterations}{An integer specifying the maximum number of iterations}

\item{ireport}{An integer specifying the report interval}
}
\value{
This function returns a list with the following entries:
\itemize{
\item par: A numeric vector with the values of the layer resistivities and thicknesses
\item value: The value or the RSS (Residual sum of squares)
\item rel.error: The value of the relative error (in percentage)
\item cal.error: A matrix with the RSS and relative error at each iteration
}
}
\description{
Function to estimate the layer parameters (real resistivities and thicknesses) from
a VES measurements using the ILPSQ (Iterative Least-Squares Procedure) with Singular
Value Decomposition proposed by Muiuane and Pedersen (1999).
}
\references{
Muiuane, E. & Pedersen, L. Automatic 1D interpretation of DC resistivity sounding
data. Journal of Applied Geophysics, 42, 1, 35-45, 1999.
}
\seealso{
Other Calibration functions: \code{\link{calibrate_svd}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
