% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ves_utilities.R
\name{convolution}
\alias{convolution}
\title{convolution}
\usage{
convolution(x, y)
}
\arguments{
\item{x}{A numeric vector of length lx}

\item{y}{A numeric vector of length ly}
}
\value{
A numeric vector of length (lx+ly-1) with the results of the convolution between the x and y vectors
}
\description{
Function to calculate the discrete convolution between two numeric vectors
}
\seealso{
Other utilities functions: \code{\link{apparent_resistivities_simple}},
  \code{\link{apparent_resistivities}},
  \code{\link{krtrans}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
