% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filt.R
\docType{data}
\name{filt}
\alias{filt}
\title{Schlumberger Filter Coefficient}
\format{A data frame with 19 rows and 1 variable:
\describe{
  \item{V1}{filter coefficient}
  ...
}}
\usage{
filt
}
\description{
Schlumberger Filter Coefficient
}
\keyword{datasets}
