% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ves_utilities.R
\name{krtrans}
\alias{krtrans}
\title{krtrans}
\usage{
krtrans(lambda, nlayers, thick, rho)
}
\arguments{
\item{lambda}{The transform parameter}

\item{nlayers}{the number of layers in the model}

\item{thick}{A numeric vector with the thicknesses of the layers in the Earth model}

\item{rho}{A numeric vector with the real resistivities of the layers in the Earth model}
}
\value{
A numeric vector with the result of the resistivity transform
}
\description{
Function to calculate the resistivity transform of a given model of a layered earth
}
\seealso{
Other utilities functions: \code{\link{apparent_resistivities_simple}},
  \code{\link{apparent_resistivities}},
  \code{\link{convolution}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
