% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing_utilities.R
\name{logseq}
\alias{logseq}
\title{logseq}
\usage{
logseq(from = 1, to = 1, n)
}
\arguments{
\item{from}{Initial point of the sequence}

\item{to}{Final point of the sequence}

\item{n}{Number of points}
}
\value{
This function returns a numeric vector with the generated sequence.
}
\description{
Function to generate sequences with logarithmic increments
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
