% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{mxad_resistivity}
\alias{mxad_resistivity}
\title{mxad_resistivity}
\usage{
mxad_resistivity(par, filter, apprho_measured, spacing)
}
\arguments{
\item{par}{A numeric vector with the values of the real resisitivity and the layer thickness}

\item{filter}{A numeric vector with the filter coefficients for a schlumberger array}

\item{apprho_measured}{A numeric vector with the values of the measured apparent resistivity}

\item{spacing}{A numeric vector with the values of AB/2.}
}
\value{
A numeric value with the maximum absolute deviation
}
\description{
Function to calculate the maximum absolute deviation between the measured and calculated apparent resistivity.
}
\seealso{
Other calibration functions: \code{\link{calibrate_nls}},
  \code{\link{calibrate}},
  \code{\link{log_mnad_resistivity}},
  \code{\link{log_mxad_resistivity}},
  \code{\link{log_rss_resistivity}},
  \code{\link{mnad_resistivity}},
  \code{\link{rss_resistivity}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
