% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ves.R
\name{plot_diagnostic}
\alias{plot_diagnostic}
\title{plot_diagnostic}
\usage{
plot_diagnostic(x)
}
\arguments{
\item{x}{A VES object, The VES must be intepreted.}
}
\value{
This function returns a ggplot2 object
}
\description{
Function to create an estimation diagnostic plot composed of:
\itemize{
\item Scatterplot between the measured and calculated apparement resistivity
\item Scatterplot between calculated apparent resistivity and the residuals
\item Scatterplot between the calculated apparent resistivity and the standarized residuals
\item QQplot of the residuals
}
}
\author{
Oscar Garcia-Cabrejo, \email{khaors@gmail.com}
}
