% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ves.R
\name{plot_transformation}
\alias{plot_transformation}
\title{plot_transformation}
\usage{
plot_transformation(x, trans.type = c("direct", "scaling", "zohdy",
  "zohdy.smoothed"))
}
\arguments{
\item{x}{A VES object}

\item{trans.type}{A character string with the type of transformation to apply. Currently only direct and
scaling, zohdy and zohdy.smoothed transformation are defined.}
}
\value{
This function returns a ggplot2 object.
}
\description{
Function to plot the resistivity-depth transformation from a given VES.
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
