% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ves.R
\name{print.ves}
\alias{print.ves}
\title{print.ves}
\usage{
\\method{print}{ves}(x, ...)
}
\arguments{
\item{x}{A VES object}

\item{...}{Additional parameters to the data.frame print function}
}
\value{
This function prints the information from the VES on the screen
}
\description{
Function to print on the screen a VES object
}
\examples{
data(ves_data1)
ab2 <- ves_data1$ab2
apprho <- ves_data1$apprho
sev1a <- ves(id= "VES1", ab2 = ab2, apprho = apprho)
print(sev1a)
}
\seealso{
Other base functions: \code{\link{plot.ves}},
  \code{\link{summary.ves}}, \code{\link{ves}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
