% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R, R/ves.R, R/ves_utilities.R
\docType{package}
\name{rves}
\alias{rves}
\alias{rves-package}
\alias{rves}
\alias{rves-package}
\alias{rves}
\alias{rves-package}
\title{rves: R package for the interpretation of Vertical Electric Soundings}
\description{
This package includes several functions used in the inversion of the VES.
}
\section{calibration functions}{


This section includes all the functions required for the estimation of the thicknesses and real resitivities from the VES data. The functions included here are:

rss_resisitivity, log_rss_resistivity, mnad_resistivity, log_mnad_resistivity, mxad_resistivity, log_mxad_resistivity, calibrate
}

\section{base functions}{

This section includes a description of the main
}

\section{utilities functions}{


This section includes all the functions required in other sections of this package.
The functions included in this section are:

convolution, krtrans, apparent_resistivities, apparent_resistivities_simple
}

