% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ves.R
\name{summary.ves}
\alias{summary.ves}
\title{summary.ves}
\usage{
\method{summary}{ves}(object, ...)
}
\arguments{
\item{object}{A VES object}

\item{...}{additional parameters for the data.frame summary function}
}
\value{
This function displays on the console a summary of the apparent resistivity data
}
\description{
Function to display a short summary of the apparent resistivity data
}
\seealso{
Other base functions: \code{\link{print.ves}},
  \code{\link{ves}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
