% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_utilities.R
\name{transform_direct}
\alias{transform_direct}
\title{transform_direct}
\usage{
transform_direct(ves, k = 2.3, alpha = 0.15)
}
\arguments{
\item{ves}{A ves object.}

\item{k}{Empirical constant.}

\item{alpha}{Decay exponent (pseudo-porosity factor).}
}
\value{
This function returns a list with the following entries:
\itemize{
\item depth: Depth
\item real.res: Real resisitivity
}
}
\description{
Function to transform the apparent resistivity into effective resistivities which are assumed closer to the real resistivities using an exponential equation.
}
\details{
The conversion from apparent resitivities to effective resistivities is achieved using the following expression:
\deqn{
\rho_{eff}=k\rho_{app}\exp{(-(1-\alpha))}
}
where:
\itemize{
\item \eqn{\rho_{eff}} is the effective resistivity which is assumed to be closer to the true resistivity.
\item \eqn{\rho_{app}} is the apparent resistivity measured in the field.
\item \eqn{k} is a empirical constant defined to improve the fit between the apparent and effective resistivities. This constant has a value of 2.3.
\item \eqn{\alpha} is the exponent in the exponential relationship. Its value is of 0.15.
}
}
\references{
Meju, M.  Simple effective resistivity-depth transformations
for infield or real-time data processing. Computers and Geosciences,
21(8), 985-992, 1995.
}
\seealso{
Other transformation functions: \code{\link{transform_scaling}},
  \code{\link{transform_smoothed_zohdy}},
  \code{\link{transform_zohdy}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
