% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_utilities.R
\name{transform_scaling}
\alias{transform_scaling}
\title{transform_scaling}
\usage{
transform_scaling(ves)
}
\arguments{
\item{ves}{A VES object}
}
\value{
This function returns a list with the following entries:
\itemize{
\item depth: Depth
\item real.res: Real resisitivity
}
}
\description{
Function to
}
\details{
The conversion from apparent to effective resistivities using the scaling approach is based on the use of the following expression:
\deqn{
\rho_{eff}^{i}=S \cdot \rho_{app}^{i}
}
where $S$ is the so called scaling factor defined as:
\deqn{
S = \frac{\rho_{app}^{i}}{ \rho_{app}^{i-1} }
}
and \eqn{\rho_{app}^{i}} and \eqn{\rho_{app}^{i-1}} are the apparent resistivities of the \eqn{i-}th and \eqn{(i-1)}-th layers respectively.
}
\references{
Meju, M.  Simple effective resistivity-depth transformations
for infield or real-time data processing. Computers and Geosciences,
21(8), 985-992, 1995.
}
\seealso{
Other transformation functions: \code{\link{transform_direct}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
