% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_utilities.R
\name{transform_smoothed_zohdy}
\alias{transform_smoothed_zohdy}
\title{transform_smoothed_zohdy}
\usage{
transform_smoothed_zohdy(ves)
}
\arguments{
\item{ves}{A VES object}
}
\value{
This function returns a list with the following entries:
\itemize{
\item depth: A numeric vector with the layer depths
\item real.res: A numeric vector with the layer resistivities
\item thickness: A numeric vector with the layer thicknesses
\item rho: A numeric vector with the layer resistivities (defined to be compatible
with calibration results)
\item rms.depth: A numeric value with the rms obtained for the depth conversion
\item rms.rho: A numeric value with the rms obtained for the resistivities.
}
}
\description{
Function used to improve the original Zohdy's method.
}
\references{
Loke, M, \& Barker, R. Improvements to the Zohdy method for the inversion of
resistivity sounding and pseudosection data. Computers \& Geosciences, 21(2),
321-332, 1995.
}
\seealso{
Other transformation functions: \code{\link{transform_direct}},
  \code{\link{transform_scaling}},
  \code{\link{transform_zohdy}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
