% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_utilities.R
\name{transform_zohdy}
\alias{transform_zohdy}
\title{transform_zohdy}
\usage{
transform_zohdy(ves)
}
\arguments{
\item{ves}{A VES object}
}
\value{
This function returns a list with the following entries:
\itemize{
\item depth: A numeric vector with the layer depths
\item real.res: A numeric vector with the layer resistivities
\item thickness: A numeric vector with the layer thicknesses
\item rho: A numeric vector with the layer resistivities (defined to be compatible
with calibration results)
\item rms.depth: A numeric value with the rms obtained for the depth conversion
\item rms.rho: A numeric value with the rms obtained for the resistivities.
}
}
\description{
Function used to obtain interpreted depths and resistivities from shifted electrode
spacings and adjusted apparent resistivities, respectively. This method does not
require optimization.
}
\references{
Zohdy, A. A new method for the automatic interpretation of Schlumberger and Wenner
sounding curves. Geopysics, 1989, 54(2), 245-253.
}
\seealso{
Other transformation functions: \code{\link{transform_direct}},
  \code{\link{transform_scaling}},
  \code{\link{transform_smoothed_zohdy}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
