% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ves_data2.R
\docType{data}
\name{ves_data2}
\alias{ves_data2}
\title{Example of a Vertical Electric Sounding}
\format{A data frame with 24 rows and 2 variable:
\describe{
  \item{ab2}{Eletrode spacing in m}
  \item{apprho}{Apparent resistivity in Ohm m}
}}
\usage{
ves_data2
}
\description{
Example of a Vertical Electric Sounding
}
\keyword{datasets}
