% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibrate_nls}
\alias{calibrate_nls}
\title{calibrate_nls}
\usage{
calibrate_nls(ves, par0, iterations = 30, ireport = 10, trace = TRUE)
}
\arguments{
\item{ves}{A ves object}

\item{par0}{A numeric vector with the values of the layer resistivities and thicknesses}

\item{iterations}{Number of iterations}

\item{ireport}{Number of iterations to report results on the console}

\item{trace}{A logical flag to indicate if the optimization information must be traced}
}
\value{
A list with the following entries:
\itemize{
\item par: A numeric vector with the values of the layer resistivities and thicknesses
\item value: The value or the RSS (Residual sum of squares)
\item rel.error: The value of the relative error (in percentage)
\item cal.error: A matrix with the RSS and relative error at each iteration
\item residual: A vector with the residuals (original scale)
\item var.residual: The variance of the residuals (log transformed scale)
\item hessian: First order approximation of the Hessian matrix. This is calculated using the Jacobian matrix.
\item cov.matrix: Covariance matrix of the estimated parameters calculated as the inverse of the
hessian matrix.
\item corr.matrix: Correlation matrix of the estimated parameters.
}
}
\description{
Function to estimate the real resistivities and thicknesses using the nonlinear least-squares approach
proposed by Roy(1999).
}
\examples{
# Example 1
data(ves_data1)
ab2 <- ves_data1$ab2
apprho <- ves_data1$apprho
sev1a <- ves(id= "VES1", ab2 = ab2, apprho = apprho)
# Let's fit a four layer model
rho <- c(40, 70, 30, 20)
thick <- c(2, 10, 50, 500)
par0 <- c(rho, thick)
res.nls1 <- calibrate_nls(sev1a, par0, iterations = 30, ireport = 5, trace = FALSE)
sev1a$rhopar <- res.nls1$rho
sev1a$thickpar <- res.nls1$thickness
sev1a$interpreted <- TRUE
plot(sev1a, type = "ves")
# Example 2
data(ves_data2)
ab2 <- ves_data2$ab2
apprho <- ves_data2$apprho
sev2a <- ves(id = "TD76", ab2 = ab2, apprho = apprho)
rho <- c(20, 50, 100, 10)
thick <- c(10, 20, 500, 100)
par0 <- c(rho, thick)
res.nls2 <- calibrate_nls(sev2a, par0, iterations = 30, ireport = 5, trace = FALSE)
sev2a$rhopar <- res.nls2$rho
sev2a$thickpar <- res.nls2$thickness
sev2a$interpreted <- TRUE
plot(sev2a, type = "ves")
}
\references{
Roy, I. An efficient non-linear least-squares 1D inversion scheme for resistivity
and IP sounding data. 1999. Geophysical Prospecting, 47, 4, 527-550.
}
\seealso{
Other calibration functions: \code{\link{calibrate_ilsqp}},
  \code{\link{calibrate_joint_nls}},
  \code{\link{calibrate_seq_joint_nls}},
  \code{\link{calibrate_step_nls}},
  \code{\link{calibrate_step}},
  \code{\link{calibrate_svd}}, \code{\link{calibrate}},
  \code{\link{log_mnad_resistivity}},
  \code{\link{log_mxad_resistivity}},
  \code{\link{log_rss_resistivity}},
  \code{\link{mnad_resistivity}},
  \code{\link{mxad_resistivity}},
  \code{\link{relative_error_resitivity}},
  \code{\link{rss_resistivity}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
