% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ves.R
\name{plot.ves}
\alias{plot.ves}
\title{plot.ves}
\usage{
\method{plot}{ves}(x, ...)
}
\arguments{
\item{x}{A VES object}

\item{...}{Additional parameters to be passed to the plot function}
}
\description{
Function to plot the Vertical Electric Sounding data. This function can create two different
types of plots: resistivity and interpretation. The resistivity plot includes the
apparent resistivity vs electrode spacing. In the interpretation plot, the values of the
apparent resistivity vs electrode spacing and the thickness and real resisitivity of the
layers are included.
}
