% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R, R/rves.R,
%   R/smoothing_utilities.R, R/transformation_utilities.R, R/ves.R,
%   R/ves_utilities.R
\docType{package}
\name{rves}
\alias{rves}
\alias{rves-package}
\alias{rves}
\alias{rves-package}
\alias{rves}
\alias{rves-package}
\alias{rves}
\alias{rves-package}
\alias{rves}
\alias{rves-package}
\alias{rves}
\alias{rves-package}
\title{rves: Interpretation of VErtical Electrical Sounding using R}
\description{
rves: Interpretation of VErtical Electrical Sounding using R

This package includes several functions used in the inversion of the VES.
}
\section{calibration functions}{


This section includes all the functions required for the estimation of the thicknesses and real resitivities from the VES data. The functions included here are:

rss_resisitivity, log_rss_resistivity, mnad_resistivity, log_mnad_resistivity, mxad_resistivity, log_mxad_resistivity, calibrate, calibrate_nls, calibrate_svd


 The calibration functions includes functions to calibrate, calculate model fit, and objective
 functions.

 The functions in this section are:

 sum_squared_residuals, mean_average_deviation, nash_sutcliffe, log_nash_sutcliffe,
 mass_balance_error, calibrate, uncertainty_quantification
}

\section{smoothing functions}{


This section includes all the functions required for the smoothing of the VES data. The functions included here are:

smoothing_ves
}

\section{Transformation functions}{


This section includes functions used to calculate the resistivity-depth transformations, The results of these functions can be used for exploratory purposes or to define an initial solution of the automatic inversion procedures. .

the functions included in this section are:

transform_direct, transform_scaling
}

\section{base functions}{

This section includes a description of the main
}

\section{utilities functions}{


This section includes all the functions required in other sections of this package.
The functions included in this section are:

convolution, krtrans, apparent_resistivities, apparent_resistivities_simple
}

