% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing_utilities.R
\name{smoothing_ves}
\alias{smoothing_ves}
\title{smoothing_ves}
\usage{
smoothing_ves(ves, method = c("smooth.spline", "kernel.regression"), ...)
}
\arguments{
\item{ves}{A Vertical Electric Sounding objet}

\item{method}{A character string specifying the smoothing method. Currently the following methods are implemented:
\itemize{
\item smooth.spline
\item kernel.regression
}}

\item{...}{Additional parameters required for the smoothing method}
}
\value{
This function returns a list with the following entries:
\itemize{
\item ab2: Values of the electrode separation
\item apprho: Values of the smoothed apparent resistivity
}
}
\description{
Function to smooth the measurements of the apparent resistivity of a VES
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
