% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ves.R
\name{ves}
\alias{ves}
\title{ves}
\usage{
ves(id = character(0), ab2 = NULL, apprho = NULL)
}
\arguments{
\item{id}{a character string defining the VES ID}

\item{ab2}{Numeric vector with the values of the electrode spacing}

\item{apprho}{Numeric vector with the values of the apparent resistivity}
}
\value{
A VES object
}
\description{
Function to create a VES object
}
\examples{
data(ves_data1)
ab2 <- ves_data1$ab2
apprho <- ves_data1$apprho
sev1a <- ves(id= "VES1", ab2 = ab2, apprho = apprho)
print(sev1a)
}
\seealso{
Other base functions: \code{\link{plot.ves}},
  \code{\link{print.ves}}, \code{\link{summary.ves}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
